feat(secret-provider): migrate to multi-tenant v2
