+++ openaev-api/src/main/java/io/openaev/rest/collector/CollectorApi.java
@@ -8,6 +8,7 @@ import io.openaev.config.TenantWriteScopeResolver;
 import io.openaev.context.TxCtx;
 import io.openaev.database.model.Action;
 import io.openaev.database.model.Collector;
+import io.openaev.database.model.ConnectorType;
 import io.openaev.database.model.ConnectorCompositeId;
 import io.openaev.database.model.ConnectorType;
 import io.openaev.database.model.ResourceType;
@@ -28,7 +29,6 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
-import java.io.IOException;
 import java.io.InputStream;
 import java.time.Instant;
 import java.util.Optional;
